Refactoring of H3M loader to make HotA format support easier
- extracted low-level reader from MapFormatH3M class
- added separate structure to define version-specific values
- cleared up some H3M format edge cases
- replaced witch hut skill vector with set
- converted several fields to enum type

+++ lib/mapping/CMapService.cpp
@@ -120,10 +120,10 @@ std::unique_ptr<IMapLoader> CMapService::getMapLoader(std::unique_ptr<CInputStre
 			case 0x00088B1F:
 				stream = std::unique_ptr<CInputStream>(new CCompressedStream(std::move(stream), true));
 				return std::unique_ptr<IMapLoader>(new CMapLoaderH3M(mapName, modName, encoding, stream.get()));
-			case EMapFormat::WOG :
-			case EMapFormat::AB  :
-			case EMapFormat::ROE :
-			case EMapFormat::SOD :
+			case static_cast<int>(EMapFormat::WOG) :
+			case static_cast<int>(EMapFormat::AB)  :
+			case static_cast<int>(EMapFormat::ROE) :
+			case static_cast<int>(EMapFormat::SOD) :
 				return std::unique_ptr<IMapLoader>(new CMapLoaderH3M(mapName, modName, encoding, stream.get()));
 			default :
 				throw std::runtime_error("Unknown map format");
